made llmcord compatible with pyinstaller again
