+++ entry_point.py
@@ -1,0 +1,3 @@
+"""Entry point for PyInstaller to create an executable"""
+
+import llmcord
